requirements.txt
fastapi~=0.100.0
uvicorn
pymongo
motor
PyJWT
python-dotenv
passlib
pydantic==2.4.2
email-validator
python-multipart
bcrypt==4.0.1